scaffold factory tests
