+++ src/factory.new.ts
@@ -1,21 +1,3 @@
-type AttributeBuilderParams<Entity, TransientParams> = {
-  sequence: number;
-  params: Partial<Entity>;
-  transientParams?: TransientParams;
-};
-
-type AttributeBuilder<
-  Entity,
-  TransientParams,
-  Property extends keyof Entity
-> = (
-  params: AttributeBuilderParams<Entity, TransientParams>
-) => Entity[Property];
-
-type EntityAttribute<Entity, TransientParams, Property extends keyof Entity> =
-  | AttributeBuilder<Entity, TransientParams, Property>
-  | Entity[Property];
-
 type EntityAttributes<Entity, TransientParams> = {
   [Property in keyof Entity]: EntityAttribute<
     Entity,
@@ -23,116 +5,156 @@ type EntityAttributes<Entity, TransientParams> = {
     Property
   >;
 };
+type EntityAttribute<Entity, TransientParams, Property extends keyof Entity> =
+  | AttributeBuilder<Entity, TransientParams, Property>
+  | Entity[Property];
+type AttributeBuilder<
+  Entity,
+  TransientParams,
+  Property extends keyof Entity
+> = (
+  params: AttributeBuilderParams<Entity, TransientParams>
+) => Entity[Property];
+type AttributeBuilderParams<Entity, TransientParams> = {
+  sequence: number;
+  params: Partial<Entity>;
+  transientParams?: TransientParams;
+};
 
-export interface EntityFactory<Entity, TransientParams, Traits> {
-  build(
-    ...args:
-      | [...Array<Traits>, Partial<Entity> & Partial<TransientParams>]
-      | Array<Traits>
+type TransientParamsForTraits<
+  Traits,
+  TraitNames extends keyof Traits
+> = UnionToIntersection<Pick<Traits, TraitNames>[TraitNames]>;
+type UnionToIntersection<U> = {
+  [K in GetKeys<U>]: U extends Record<K, infer T> ? T : never;
+};
+type GetKeys<U> = U extends Record<infer K, any> ? K : never;
+
+type FinalBuilder<Builder> = Omit<Builder, 'attributes' | 'transient'>;
+
+interface TraitBuilder<
+  Entity,
+  GlobalTransientParams = unknown,
+  TransientParams = {}
+> {
+  transient<TraitTransientParams extends Record<string, unknown>>(
+    params: TraitTransientParams
+  ): Omit<
+    TraitBuilder<
+      Entity,
+      GlobalTransientParams,
+      TransientParams & TraitTransientParams
+    >,
+    'transient'
+  >;
+
+  attributes(
+    attributes: EntityAttributes<
+      Partial<Entity>,
+      GlobalTransientParams & TransientParams
+    >
+  ): FinalBuilder<TraitBuilder<Entity, GlobalTransientParams, TransientParams>>;
+
+  afterCreate(
+    afterCreateCallback: (
+      entity: Entity,
+      params: { transientParams: GlobalTransientParams & TransientParams }
+    ) => void
+  ): FinalBuilder<TraitBuilder<Entity, GlobalTransientParams, TransientParams>>;
+}
+
+interface FactoryBuilder<
+  Entity = unknown,
+  GlobalTransientParams = unknown,
+  Traits = unknown
+> {
+  transient<GlobalTransientParams>(
+    params: GlobalTransientParams
+  ): Omit<FactoryBuilder<Entity, GlobalTransientParams, Traits>, 'transient'>;
+
+  attributes<Entity>(
+    attributes: EntityAttributes<Entity, GlobalTransientParams>
+  ): FinalBuilder<FactoryBuilder<Entity, GlobalTransientParams, Traits>>;
+
+  trait<Trait extends string, TraitTransientParams>(
+    name: Trait,
+    traitBuilderCallBack: (
+      builder: TraitBuilder<Entity, GlobalTransientParams>
+    ) => FinalBuilder<
+      TraitBuilder<Entity, GlobalTransientParams, TraitTransientParams>
+    >
+  ): FinalBuilder<
+    FactoryBuilder<
+      Entity,
+      GlobalTransientParams,
+      Traits & Record<Trait, TraitTransientParams>
+    >
+  >;
+
+  afterCreate(
+    afterCreateCallback: (
+      entity: Entity,
+      params: { transientParams: GlobalTransientParams }
+    ) => void
+  ): FinalBuilder<FactoryBuilder<Entity, GlobalTransientParams, Traits>>;
+}
+
+interface EntityFactory<Entity, GlobalTransientParams, Traits> {
+  build(entity?: Partial<Entity & GlobalTransientParams>): void;
+
+  build<TraitNames extends keyof Traits>(
+    ...params:
+      | [
+          ...traits: [TraitNames, ...Array<TraitNames>],
+          params: Partial<
+            Entity &
+              GlobalTransientParams &
+              TransientParamsForTraits<Traits, TraitNames>
+          >
+        ]
+      | Array<keyof Traits>
   ): void;
 }
 
-export interface FactoryBuilder<Entity, TransientParams> {
-  attributes(
-    attributes: EntityAttributes<Entity, TransientParams>
-  ): FactoryBuilder<Entity, TransientParams>;
-  trait(
-    name: string,
-    builderCallback: (builder: TraitBuilder<Entity, TransientParams>) => void
-  ): FactoryBuilder<Entity, TransientParams>;
-  afterCreate(
-    afterCreateCallback: (
-      entity: Entity,
-      params: { transientParams: TransientParams }
-    ) => void
-  ): FactoryBuilder<Entity, TransientParams>;
-}
-
-export interface TraitBuilder<Entity, TransientParams> {
-  attributes(
-    attributes: EntityAttributes<Partial<Entity>, TransientParams>
-  ): TraitBuilder<Entity, TransientParams>;
-  afterCreate(
-    afterCreateCallback: (
-      entity: Entity,
-      params: { transientParams: TransientParams }
-    ) => void
-  ): TraitBuilder<Entity, TransientParams>;
-}
-
 export const Factory = {
-  define<Entity, TransientParams = Record<string, any>, Traits = string>(
+  define<Entity, GlobalTransientParams, Traits>(
     factoryBuilderCallback: (
-      builder: FactoryBuilder<Entity, TransientParams>
-    ) => void
-  ) {
-    const factory: EntityFactory<Entity, TransientParams, Traits> = {
-      build(...args) {},
+      builder: FactoryBuilder
+    ) => FinalBuilder<FactoryBuilder<Entity, GlobalTransientParams, Traits>>
+  ): EntityFactory<Entity, GlobalTransientParams, Traits> {
+    const factory: EntityFactory<Entity, GlobalTransientParams, Traits> = {
+      build(...params: Array<any>) {},
     };
 
-    const builder: FactoryBuilder<Entity, TransientParams> = {
+    const factoryBuilder: FactoryBuilder<any, any, any> = {
       attributes(attributes) {
-        return builder;
+        return factoryBuilder;
       },
-      trait(name, traitBuilderCallback) {
-        const traitBuilder: TraitBuilder<Entity, TransientParams> = {
+      transient(params) {
+        return factoryBuilder;
+      },
+      trait(name, traitBuilderCallBack) {
+        const traitBuilder: TraitBuilder<any, any, any> = {
           attributes(attributes) {
+            return traitBuilder;
+          },
+          transient(params) {
             return traitBuilder;
           },
           afterCreate(afterCreateCallback) {
             return traitBuilder;
           },
         };
-        traitBuilderCallback(traitBuilder);
-        return builder;
+        traitBuilderCallBack(traitBuilder);
+        return factoryBuilder;
       },
       afterCreate(afterCreateCallback) {
-        return builder;
+        return factoryBuilder;
       },
     };
 
-    factoryBuilderCallback(builder);
+    factoryBuilderCallback(factoryBuilder);
 
     return factory;
   },
 };
-
-interface User {
-  id: number;
-  firstName: string;
-  lastName: string;
-  isAdmin: boolean;
-}
-
-type UserTransientParams = { foo: boolean; bar: string };
-type UserTraits = 'admin' | 'inactive';
-const userFactory = Factory.define<User, UserTransientParams, UserTraits>(
-  (factory) =>
-    factory
-      .attributes({
-        id: 1,
-        firstName: 'foo',
-        lastName({ sequence, transientParams }) {
-          return transientParams?.foo ? `last${sequence}` : 'last';
-        },
-        isAdmin: false,
-      })
-      .trait('admin', (trait) =>
-        trait
-          .attributes({
-            isAdmin: true,
-          })
-          .afterCreate((entity, { transientParams }) => {
-            console.log('trait after create', entity, transientParams);
-          })
-      )
-      .afterCreate((entity, { transientParams }) => {
-        console.log('factory after create', entity, transientParams);
-      })
-);
-
-userFactory.build({ firstName: 'foo' });
-userFactory.build('admin', { firstName: 'foo' });
-userFactory.build('admin', 'inactive', { firstName: 'bar' });
-
-
